pyproject.toml
[project]
name = "flask-boilerplate"
version = "0.1.0"
description = "Contains all dependencies that will be used for creating an API application with Flask."
authors = [
    {name = "Mark Achiles",email = "markachilesflores2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "apispec (>=6.8.1,<7.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "marshmallow-sqlalchemy (>=1.4.1,<2.0.0)",
    "flask-smorest (>=0.45.0,<0.46.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "marshmallow-enum (>=1.5.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "flask-cors (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
